Readme
Catch the Runner is a fun android game. Although the gameplay is simple, scoring is not easy. This feature makes the game more ambitious.

Road Runner appears periodically on the screen. Score increases whenever it is touched. We try to get the highest score. There are 3 difficulty levels; easy, medium and hard. The highest score is also available on home screen, as well as each screen.